코드 수정

+++ config/views.py
@@ -1,3 +1,4 @@
+import http
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
 from django.shortcuts import render
@@ -36,6 +37,7 @@ def signin(request):
       # select * from user where email=? and pwd=?
       user = User.objects.get(email=email, pwd=pwd)
       request.session['email'] = email
+      request.session['name'] = user.name
       return render(request, 'signin_success.html')
     except:
       return render(request, 'signin_fail.html')
@@ -68,11 +70,33 @@ def write(request):
 
   return render(request, 'write.html')
 
+from django.core.paginator import Paginator
+
 def list(request):
+  page = request.GET.get('page')
+
   # select * from article order by id desc
   article_list = Article.objects.order_by('-id')
+
+  p = Paginator(article_list, 10)
+  try:
+    page = int(page)
+    article_list = p.page(page)
+  except:
+    page = 1
+    article_list = p.page(page)
+
+# 페이지네이션 중 시작페이지
+  start_page = (page - 1) // 10 * 10 + 1
+# 페이지네이션 중 마지막페이지
+  end_page = start_page + 9
+# 전체 페이지 수가 end_page 보다 적다면...
+  if end_page > p.num_pages:
+    end_page = p.num_pages
+
   context = { 
-    'article_list' : article_list 
+    'article_list' : article_list, 
+    'page_info' : range(start_page, end_page +1)
   }
   return render(request, 'list.html', context)
 
@@ -88,6 +112,17 @@ def update(request, id):
   # select * from article where id = ?
   article = Article.objects.get(id=id)
 
+  # 로그인한 사용자의 정보 확인
+  name = request.session.get('name')
+  # 작성자명 확인
+  if article.user.name != name:
+     return HttpResponse('''
+      <script>
+        alert("작성자만 수정 가능합니다.");
+        location = "/article/detail/%s/";
+      </script>
+    ''' % id)
+
   if request.method == 'POST':
     title = request.POST.get('title')
     content = request.POST.get('content')
@@ -99,7 +134,7 @@ def update(request, id):
       article.save()
       return render(request, 'update_success.html')
     except:
-      return render(request, 'update_fail.html')
+        return render(request, 'update_fail.html')
 
   context = { 
     'article' : article 
@@ -109,8 +144,20 @@ def update(request, id):
 def delete(request, id):
   try:
     # select * from article where id = ?
+    name = request.session['name']
     article = Article.objects.get(id=id)
-    article.delete()
+
+    if article.user.name == name:
+      article.delete()
+    else:
+
+      return HttpResponse('''
+      <script>
+      alert("작성자만 삭제 가능합니다.");
+      location = "/article/detail/%s/";
+      </script>
+    ''' % id)
+
     return render(request, 'delete_success.html')
   except:
     return render(request, 'delete_fail.html')
